# openApi3.yml
openapi: 3.0.0
info:
  title: fCount API
  version: 1.0.0
  description: This is the API for fCount, a service to count human faces in images
servers:
  - url: 'http://localhost:3100'
    description: Local dev server
paths:
  /f-counts:
    post:
      summary: Create a new fCount request to count faces in an image
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/fCountRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      callbacks:
        sendFCount:
          '{$request.body#/callbackUrl}':
            post:
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/fCountResponse'
              responses:
                default:
                  description: Unknown -- may depend on the service
    get:
      summary: Get a list of fCount requests
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/fCountResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /f-counts/{id}:
    get:
      summary: Get the status of a fCount request with results if available
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fCountResponse'
        '404':
          description: Not Found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a fCount request
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    fCountRequest:
      type: object
      properties:
        imgUrl:
          type: string
          format: uri
        callbackUrl:
          type: string
          format: uri
    fCountResponse:
      type: object
      properties:
        id:
          type: number 
        url:
          type: string
          format: uri
        status:
          type: string
        fcount:
          type: number
          nullable: true 